chore(coupon-api): move in domain package
