+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/coupon/domain/CouponTypeConverter.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.api.coupon.application;
+package dev.be.coupon.api.coupon.domain;
 
 import deb.be.coupon.CouponType;
 import dev.be.coupon.api.coupon.domain.exception.InvalidCouponTypeException;
@@ -7,7 +7,7 @@ public class CouponTypeConverter {
 
     public static CouponType from(final String type) {
         if (type == null || type.isBlank()) {
-            throw new InvalidCouponTypeException("쿠폰 타입이 null이거나 비어있습니다.");
+            throw new InvalidCouponTypeException("쿠폰을 생성하려면 유효한 쿠폰 타입이 존재해야 합니다.");
         }
 
         try {
